mmap reading implemented and tested. write buffer and async flushing implemented and tested. 10^5 pages 10^3 words per page invert index calculating time: 3 min 40 sec on MacBook Air 2012

+++ mapreduce/core/InterResultReader.h
@@ -17,7 +17,7 @@ public:
 	InterResultLoader(std::string filename, EmitDumper *dumper);
 	~InterResultLoader();
 	
-	EmitType *readEmit(size_t offset);
+	EmitType *readEmit(off_t offset);
 };
 
 #endif
